% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_wpopbuilding.R
\name{combine_wpopbuilding}
\alias{combine_wpopbuilding}
\title{Combine WORLDPOP building tifs data within a shapefile}
\usage{
combine_wpopbuilding(
  ldrive = "./../S2S-REMDI/GIN_2021",
  identifier = "buildings_v2_0",
  shp = "./../S2S-REMDI/GIN_2021/sous_prefectures.shp"
)
}
\description{
This function combines the data obtained from World Pop database on buildings with the filenames structured like
XXX_buildings_v2_0_indicator within a specified shapefile
}
