% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_extramergevars.R
\name{drop_extramergevars}
\alias{drop_extramergevars}
\title{Clean up data by dropping variables systematically}
\usage{
drop_extramergevars(dt, string_match = "i\\\\.")
}
\arguments{
\item{dt}{data.table or data.frame object}

\item{string_match}{a string match to identify the set of additional variables to be dropped. "i\\."
will drop all variables that contain duplicated variables from a data.table merge}
}
\description{
This simple helper function will drop duplicated variable names created via data.table merges
}
