% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdi_maketables.R
\name{emdi_maketables}
\alias{emdi_maketables}
\title{Create report suitable tables from EMDI-EBP function outputs}
\usage{
emdi_maketables(obj = ginemdi_model2, outfolder = "data")
}
\arguments{
\item{obj}{object of classes 'ebp', 'emdi' i.e. the resulting object of the ebp survey to survey function (emdi::ebp())}

\item{docx}{if TRUE, the tables will be outputted straight to Microsoft Word (Ensure Java installation on your system to use
this option)}
}
\description{
Create report suitable tables from EMDI-EBP function outputs
}
