% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_datapull2.R
\name{gee_datapull2}
\alias{gee_datapull2}
\title{Pull spatial data from google earth engine (GEE) for a given region on any feature in the google earth engine database}
\usage{
gee_datapull2(
  email = "ifeanyi.edochie@gmail.com",
  gee_polygons = "users/ifeanyiedochie/cmr_polypop_boundary",
  gee_dataname = "NOAA/VIIRS/DNB/MONTHLY_V1/VCMSLCFG",
  gee_datestart = "2018-01-01",
  gee_dateend = "2018-01-31",
  gee_band = "avg_rad",
  gee_scale = 100,
  gee_desc = "nighttimelight_cmr",
  gee_stat = "mean",
  gdrive_folder = "/SAEplus",
  ldrive_dsn = "data/cmr_nighttimelight",
  gee_crs = "EPSG:4326"
)
}
\arguments{
\item{email}{A GEE authenticated gmail address}

\item{gee_dataname}{The specific google earth engine collection dataset name (use default as example)
see (https://developers.google.com/earth-engine/datasets/) for full name}

\item{gee_datestart}{The starting date for the specific feature of interest}

\item{gee_dateend}{The ending date for the specific feature of interest}

\item{gee_band}{The bandname within the gee_dataname selected (see https://developers.google.com/earth-engine/datasets/)}

\item{gee_desc}{The name to be used to name output in google drive as well as local drive}

\item{gdrive_folder}{Google Drive folder name to be created or uses if already existing to store shapefile output}

\item{ldrive_dsn}{Full file path (including shapefile name) for local storage of resulting shapefile}

\item{gee_crs}{Set CRS}

\item{gee_boundary}{A shapefile with the boundaries for a specific region/country for which data is to be collected}

\item{gee_image}{if TRUE, an image rather than an image collection is expected (FALSE by default)}

\item{scale}{Used in mean region reduction for zonal statistics}
}
\value{
shapefiles to local drive
}
\description{
Pull spatial data from google earth engine (GEE) for a given region on any feature in the google earth engine database
}
