% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_pullbigimage.R
\name{gee_downloadimage}
\alias{gee_downloadimage}
\title{Pull Large Images from Google Earth Engine's Image Collections for further extraction}
\usage{
gee_downloadimage(
  gee_country,
  gee_dataname,
  gee_band,
  gee_scale = NULL,
  gdrive_folder = "rgee",
  ldrive_dsn,
  gee_datestart,
  gee_dateend,
  gee_desc,
  gee_crs = NULL
)
}
\arguments{
\item{gee_dataname}{The specific google earth engine collection dataset name (use default as example)
see (https://developers.google.com/earth-engine/datasets/) for full name}

\item{gee_band}{The bandname within the gee_dataname selected (see https://developers.google.com/earth-engine/datasets/)}

\item{gdrive_folder}{Google Drive folder name to be created or uses if already existing to store shapefile output}

\item{ldrive_dsn}{Full file path (including shapefile name) for local storage of resulting shapefile}

\item{gee_desc}{The name to be used to name output in google drive as well as local drive}

\item{email}{A GEE authenticated gmail address}

\item{gee_polygon}{A polygon shapefile spanning the region in the gee_boundary file for features will be computed.
forms the unit of analysis}

\item{scale}{Used in mean region reduction for zonal statistics}
}
\value{
shapefiles to local drive
}
\description{
This function is a slight variation from gee_datapull() which pulls image collections. This function specifically pulls
images
}
