% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_pullbigdata.R
\name{gee_pullbigdata}
\alias{gee_pullbigdata}
\title{Another pulling function for pulling Google Earth Engine data for larger maps}
\usage{
gee_pullbigdata(
  email = "ifeanyi.edochie@gmail.com",
  shp_dsn,
  shp_layer,
  gee_name = "NOAA/VIIRS/DNB/MONTHLY_V1/VCMCFG",
  gee_datestart = "2019-01-01",
  gee_dateend = "2019-12-31",
  gee_band = "avg_rad",
  gee_chunksize = 4000,
  gee_stat = "mean",
  gee_scale = 1000
)
}
\arguments{
\item{email}{A GEE authenticated gmail address}

\item{shp_dsn}{character of length 1; the local folder location of the shapefile}

\item{shp_layer}{character of length 1; the name of the file}

\item{gee_datestart}{The starting date for the specific feature of interest}

\item{gee_dateend}{The ending date for the specific feature of interest}

\item{gee_chunksize}{an integer; the size of each chunk which the shapefile should be divided into for faster
pulls (GEE collection query aborts after accumulating a chunk size with over 5000 elements i.e.
rows and columns combined)}

\item{gee_stat}{the zonal statistic to be estimated. The current options are mean, median, min, max and stdDev}

\item{gee_dataname}{character of length 1; The specific google earth engine collection dataset name}
}
\value{
an object of class data.frame
}
\description{
Another pulling function for pulling Google Earth Engine data for larger maps
}
