% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_pullimage.R
\name{gee_pullimage}
\alias{gee_pullimage}
\title{Pull Image Data from Google Earth Engine}
\usage{
gee_pullimage(
  email = "ifeanyi.edochie@gmail.com",
  gee_polygons = "users/ifeanyiedochie/cmr_polypop_boundary",
  gee_dataname = "Tsinghua/FROM-GLC/GAIA/v10",
  gee_band = "change_year_index",
  scale = 30,
  gee_desc = "impervious_cmr",
  gee_stat = "mean",
  gdrive_folder = "/SAEplus",
  ldrive_dsn = "data/cmr_impervious"
)
}
\description{
This function is also sends data pull requests to the GEE server specifically for Image tags as
different from 'SAEplus::gee_datapull()' which pulls Image Collection tags specifically
}
