% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gengrid.R
\name{gengrid}
\alias{gengrid}
\title{Polygonize a raster and compute zonal statistics at polygon and full shape-file level}
\usage{
gengrid(
  dsn = "data-raw",
  layer = "gadm36_CMR_0",
  stats = "sum",
  featname = "population",
  raster_tif = "cmr_ppp_2020_UNadj_constrained.tif",
  grid_shp = T,
  grid_size = 1,
  crs = "+proj=longlat +datum=WGS84 +no_defs",
  drop_Zero = T
)
}
\arguments{
\item{dsn}{stands for data source name (see sf::st_read documentation for more information)}

\item{layer}{layer name (see sf::st_read documentation for more information), will also accept object}

\item{stats}{zonal statistics to be estimated}

\item{featname}{the feature that the zonal statistic is computed for}

\item{raster_tif}{raster file with tif extension}

\item{grid_shp}{if TRUE, a grid system will be created for the shapefile (shp), grid_size must be specified}

\item{grid_size}{the diagonal length of the polygon in km}

\item{crs}{the co-ordinate reference system to be used}

\item{drop_zero}{if TRUE, gengrid will keep only zonal statistics that are different from zero}
}
\value{
list object aggregated zonal statistic, polygon data and summary statistics on polygon data
}
\description{
Polygonize a raster and compute zonal statistics at polygon and full shape-file level
}
\examples{

}
