% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_pull.R
\name{hdx_pull}
\alias{hdx_pull}
\title{Check and pull data from Humanitarian Data Exchange}
\usage{
hdx_pull(
  iso = "BEN",
  identifier = "Relative wealth indicator",
  location_folder = NULL
)
}
\arguments{
\item{iso}{The standard ISO-3 country code for country of interest}

\item{identifier}{Type of data to check}

\item{location_folder}{Destination for the downloaded files}
}
\description{
Check and pull data from Humanitarian Data Exchange
}
