% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Select Model for Small Area Estimation}
\usage{
model_selection(dt, outcomevar, dropvars)
}
\arguments{
\item{dt}{object of class data.table or data.frame}

\item{outcomevar}{character vector of size 1 naming the outcome variable}

\item{dropvars}{a list of variables to be omitted by the selection algorithm, all other variables will be included}
}
\value{
a list summarizing the lasso regression output as well as the set of coefficients
}
\description{
A function to select the set of geospatial variables to be used for a household level small area estimation model
}
