% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_gengrid.R
\name{mult_gengrid}
\alias{mult_gengrid}
\title{Polygonize multiple rasters (TIFs) and compute zonal statistics at polygon and full shape-file level}
\usage{
mult_gengrid(
  tif_namelist = gin_buildingtifs,
  location = "tests/testdata",
  feature_name = c("count", "cv_length", "imagery_year", "mean_length", "total_length",
    "cv_area", "density", "mean_area", "total_area", "urban"),
  parallel = T,
  numCores = length(gin_buildingtifs)
)
}
\arguments{
\item{tif_namelist}{a list object of raw tif names to be extracted}

\item{location}{a folder containing the tif files}

\item{feature_name}{a list of names for each element of tif_namelist to name the indicator to be extracted}

\item{parallel}{if TRUE, parallelize the process of extract files to speed the gengrid() process}

\item{numCores}{the number of server cores to be used for this}
}
\description{
This function builds on the gengrid() function to extract data from a list of
tagged image files.
}
