% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_cleanprocess.R
\name{osm_cleanprocess}
\alias{osm_cleanprocess}
\title{This function does simple cleaning of the OSM data for lines, multipolygons and points}
\usage{
osm_cleanprocess(
  dt,
  dropNA_cols = TRUE,
  zero_NA = TRUE,
  missing_threshold = 0.3
)
}
\arguments{
\item{dropNA_cols}{if TRUE, unnamed classifications of features are dropped. Note: This is different from the
label "unclassified" as some features are tagged "NA" in OSM data (default is FALSE)}

\item{zero_NA}{if TRUE, NA observations are left as zero (default is FALSE)}

\item{missing_threshold}{numeric, if zero_NA is TRUE, missing_threshold is a value between 0 and 1. Columns with a rate
of missing observations less than missing_threshold will have missing values converted to 0.}
}
\value{
a cleaned data.table object
}
\description{
This function does simple cleaning of the OSM data for lines, multipolygons and points
}
