% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_datapull.R
\name{osm_datapull}
\alias{osm_datapull}
\title{This function pulls specific layer data (lines, multipolygons and points) from the Open Street Maps for countries or regions}
\usage{
osm_datapull(
  country = "Cameroon",
  ldrive = "C:/Users/ifean/Documents/WorldBankWork/SAEPlus_Other",
  lines = T,
  points = T,
  multipolygon = T
)
}
\arguments{
\item{country}{a string argument for the country name (or city) of interest i.e. place argument for oe_get function}

\item{ldrive}{a string argument for the local folder/drive location where the data should be stored}

\item{lines}{if TRUE, OSM features with lines data will be pulled}

\item{points}{if TRUE, OSM features with points data will be pulled}

\item{multipolygon}{if TRUE, OSM features will be pulled within polygons}
}
\description{
This function pulls specific layer data (lines, multipolygons and points) from the Open Street Maps for countries or regions
}
