% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_processlines.R
\name{osm_processlines}
\alias{osm_processlines}
\title{This function creates indicators from OSM lines data in preparation for survey to survey imputation analysis}
\usage{
osm_processlines(
  shapefile_path = "data/cmr_polypop_boundary.shp",
  osm_path = "C:/Users/ifean/Documents/WorldBankWork/SAEPlus_Other/Cameroon_osmlines",
  geoid_var = "id",
  feature_var = "highway"
)
}
\arguments{
\item{shapefile_path}{filepath of country/region shapefile with polygons of interest}

\item{osm_path}{filepath to open street map lines data (road networks)}

\item{geoid_var}{the variable that points to the common identifier ID between shapefile_path object and osm_path object}

\item{feature_var}{specific feature of interest in the osm_path object}
}
\value{
A list containing two objects: OSM lines data.table in both wide and long formats
}
\description{
This function creates indicators from OSM lines data in preparation for survey to survey imputation analysis
}
