% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_processmp.R
\name{osm_processmp}
\alias{osm_processmp}
\title{This function computes the number of points for the multipolygon data based on selected features}
\usage{
osm_processmp(
  shapefile_path = "data/cmr_polypop_boundary.shp",
  osm_path = "C:/Users/ifean/Documents/WorldBankWork/SAEPlus_Other/Cameroon_osmmp",
  geoid_var = "id",
  feature_var = c("amenity", "office", "shop")
)
}
\arguments{
\item{shapefile_path}{filepath of country/region shapefile with polygons of interest}

\item{osm_path}{filepath to open street map lines data (road networks)}

\item{geoid_var}{the variable that points to the common identifier ID between shapefile_path object and osm_path object}

\item{feature_var}{specific feature of interest in the osm_path object}
}
\value{
A list containing two objects: a dataframe/datatable with original OSM multipolygon data containing feature length
in polygon shapefile (from shapefile_path argument) and a lazy datatable version of the full OSM multipolygon data
with point information
}
\description{
This function computes the number of points for the multipolygon data based on selected features
}
