% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_rasterreduceregions.R
\name{parallel_extract}
\alias{parallel_extract}
\title{Extract group of rasters into the same grid}
\usage{
parallel_extract(shp_dt, raster_list, fun_list, numCores)
}
\arguments{
\item{shp_dt}{the target shapefile to be extracted into}

\item{raster_list}{the list of rasters to be extracted}

\item{raster_fun}{the raster function}
}
\description{
This function takes a list of rasters (rasters of the same theme, E.g daily rainfall, daily nighttimelights) and
computes a statistic which is extracted into the shapefile
}
