% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_addbenchmark.R
\name{saeplus_addbenchmark}
\alias{saeplus_addbenchmark}
\title{A simple function to include benchmarked results to preferred file}
\usage{
saeplus_addbenchmark(
  ebp_obj = ginemdi_model2,
  csv_file = "data/emdi_results2.csv"
)
}
\arguments{
\item{ebp_obj}{an object of class "ebp"}

\item{csv_file}{full file name (including path) where results will be outputted}

\item{bm_obj}{an object of class data.table / data.frame which is the result of benchmarking the poverty rates
(usually the output of the function saeplus_calibratepovrate())}

\item{bm_unitid}{the id variable within the bm_obj object argument}
}
\description{
This function is built to take the results of the benchmarked poverty rates and use them to replace
the original ebp estimates that were written to an xlsx file using the SAEplus::emdi_writeexcel()
or emdi::write.excel() functions
}
