% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_calibratepovrate.R
\name{saeplus_calibratepovrate}
\alias{saeplus_calibratepovrate}
\title{This function benchmarks the poverty rates to match survey estimates}
\usage{
saeplus_calibratepovrate(
  hh_dt = hh.dt,
  pop_dt = gin_masterpoly.dt,
  ebp_obj = ginemdi_model2,
  area_id = "ADM3_CODE",
  pop_var = "population",
  harea_id = "ADM1_CODE",
  povline = 5006362,
  welfare = "pcexp",
  weight = "hhweight",
  excl_outsample = TRUE
)
}
\arguments{
\item{hh_dt}{a household data.frame/data.table object containing area level IDs for which poverty rates were small area
estimated}

\item{pop_dt}{a population level data.frame/data.table object containing geospatial polygons and areas level IDs used in hh_dt
for which population level estimates can be computed}

\item{ebp_obj}{an object of class "emdi" "ebp" obtained from small area estimation containing area level poverty rates.
Area Level IDs must be matching across pop_dt, hh_dt and ebp_obj}

\item{area_id}{the level at which the poverty areas were estimated}

\item{harea_id}{a higher level at which intermediate SAE estimates will be computed}

\item{povline}{national poverty line to be used with the hh_dt object to determine survey level poverty rates}

\item{weight}{household level weights within the hh_dt object}

\item{excl_outsample}{whether or not outsample grids should be excluded in the benchmarking exercise}
}
\description{
This function ensures that the estimated poverty rates obtained from aggregating the predicted small area estimated
matches with the overall poverty rates obtained from the general survey
}
