% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_checkshpfiles.R
\name{saeplus_checkshpfiles}
\alias{saeplus_checkshpfiles}
\title{Check for inconsistencies in the administrative boundaries shapefile}
\usage{
saeplus_checkshpfiles(shp_dt, lowest_admin, crs = 4326, file)
}
\arguments{
\item{shp_dt}{a shapefile dataframe (class sf, data.frame/data.table) object}

\item{lowest_admin}{a variable name for the lowest admin level (typically admin 3)}

\item{crs}{the coordinate reference system associated with the shp_dt}

\item{file}{the filename (including extension) to which a geospatial plot to show duplicated regions}
}
\description{
This function checks for errors that are often common in admin boundary shapefiles and could lead to
potential problems when joining admin boundary shapefile data with other datasets such as geospatial
household surveys
}
