% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_makesumtable.R
\name{saeplus_conssummary}
\alias{saeplus_conssummary}
\title{Compute Outcome Variable Summary Statistics}
\usage{
saeplus_conssummary(
  dt = gin_master.dt,
  out_var = "pcexp",
  weight = "popweight",
  group_var = "ADM1_NAME",
  threshold = 5006362,
  hhid = "hhid"
)
}
\arguments{
\item{dt}{a data.table or data.frame object (object will be coerced to data.table for analysis)}

\item{out_var}{a character variable for outcome variable of interest (for instance household consumption estimate)}

\item{weight}{a weighting variable (replace with vector of 1s if not including weights)}

\item{group_var}{a grouping variable}

\item{threshold}{threshold to be applied to out_var (for example a poverty line)}

\item{hhid}{household ID variable}
}
\value{
a table
}
\description{
A function to compute summary statistics for an outcome variable at the group level
}
