% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_dropchars.R
\name{saeplus_dropchars}
\alias{saeplus_dropchars}
\title{Remove non-numeric characters}
\usage{
saeplus_dropchars(dt, vars)
}
\arguments{
\item{dt}{a data.table/data.frame/matrix object}

\item{vars}{one or more variables from which non-characters will be removed}
}
\description{
This function removes non-numeric characters from one or more specified columns within data.frame/data.table/matrix
objects. Then the results will be converted to integers
}
