% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_dummify.R
\name{saeplus_dummify}
\alias{saeplus_dummify}
\title{Create dummy variables}
\usage{
saeplus_dummify(dt, var)
}
\arguments{
\item{dt}{an object of class data.table/data.frame}

\item{var}{a variable name within dt}
}
\description{
This function will create a set of dummy variables from the level of another variable
}
