% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_gencensus.R
\name{saeplus_gencensus}
\alias{saeplus_gencensus}
\title{Create a synthetic census}
\usage{
saeplus_gencensus(
  poly_dt = polyest,
  hhsize = "ind_estimate",
  popsize = "population",
  poly_id = "id"
)
}
\arguments{
\item{poly_dt}{sf or dataframe object with polygons or multipolygons with average household size included}

\item{hhsize}{string for the name of the average household size variable}

\item{popsize}{string for the name of the population size for each polygon}

\item{poly_id}{string for the polygon id variable to be used for expanding poly_dt}
}
\description{
A function that generates a synthetic census from household level survey data with outcome and covariates and merges this
data with household survey data for
}
