% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_hhestpoly.R
\name{saeplus_hhestpoly}
\alias{saeplus_hhestpoly}
\title{Compute the geospatial grid level estimates}
\usage{
saeplus_hhestpoly(
  geo_dt,
  hh_dt,
  shp_dt,
  admin_var = "ADM3_CODE",
  ind_var = "hhsize",
  weight_var = "hhweight",
  FUN = "weighted.mean"
)
}
\arguments{
\item{geo_dt}{sf/st dataframe object including the polygon geometry included}

\item{hh_dt}{household survey data with indicator(s)}

\item{shp_dt}{shapefile with desired administrative level of aggregation}

\item{admin_var}{a variable within hh_dt for the level at which the indicator will be estimated}

\item{ind_var}{the indicator variable of interest within hh_dt}

\item{weight_var}{the weight variable of interest within hh_dt}

\item{FUN}{the aggregation function to be applied (currently MUST take ind_var and weight_var)}

\item{geo_dt_var}{existing geospatial indicator (such as population) to be used in estimating grid level estimates}
}
\value{
data.table/data.frame object with grid level estimates
}
\description{
This function estimates geospatial grid level estimates from household surveys indicators
}
