% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_modelsubarea.R
\name{saeplus_modelsubarea}
\alias{saeplus_modelsubarea}
\title{Small Estimation of a Sub Area Level Model}
\usage{
saeplus_modelsubarea(
  hhsurvey_dt,
  geopolycensus_dt,
  geopoly_id,
  size_hh,
  geopopvar,
  wgt_vartype,
  weight,
  cons_var,
  pline,
  cand_vars
)
}
\arguments{
\item{hhsurvey_dt}{an object of class "sf" and "data.table/data.frame" representing the household unit level data with
welfare variable}

\item{geopolycensus_dt}{an object of class sf, data.table and/or data.frame containing polygon/multipolygon
geometries and geospatial indicators}

\item{geopoly_id}{a string/character variable representing the polygon ID within geopolycensus_dt}

\item{size_hh}{an integer/numeric for household size variable within the hhsurvey_dt object}

\item{geopopvar}{a character string for the population count variable name in the geopolycensus_dt}

\item{wgt_vartype}{a character string representing the weighting type in hhsurvey_dt. The options
could be "hh", "pop" i.e. households vs population weights.}

\item{weight}{a numeric/integer weight variable found within the hhsurvey_dt}

\item{cons_var}{the dependent variable for small area estimation (typically household per capita consumption)}

\item{pline}{the national poverty line (ensure the same units as cons_var)}

\item{cand_vars}{a character vector of candidate explanatory variables to be included in the model
selection process}

\item{target_id}{a character string representing an integer column vector for the admin level at which
small area estimates will be computed for the poverty map}

\item{ncpu}{the number of CPUs for parallelizing the small area estimation algorithm}
}
\description{
This function takes as inputs the geospatial census of remote sensing data, a geocoded household survey with welfare
aggregates and then estimates the empirical best predictor of the poverty rate.
}
