% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_modelunitlevel.R
\name{saeplus_modelunitlevel}
\alias{saeplus_modelunitlevel}
\title{Estimate the unit level model}
\usage{
saeplus_modelunitlevel(
  hhsurvey_dt,
  hhsurvey_lat,
  hhsurvey_lon,
  hhid_var,
  hhsize,
  adminshp_dt,
  hhgeocodes_dt = NULL,
  geopolycensus_dt,
  geopopvar,
  crs_set = rep(4326, 3),
  agr_set = rep("constant", 3),
  cand_vars,
  outcome_var,
  wgt_vartype = "hh",
  weight,
  create_dummy = TRUE,
  dummy_var
)
}
\arguments{
\item{hhsurvey_dt}{object of class data.frame/data.table corresponding to household survey data (unit level data)}

\item{hhsurvey_lat}{the latitude variable within the hh_dt object}

\item{hhsurvey_lon}{the longitude variable within the hh_dt object}

\item{hhid_var}{the household ID variable from the hhsurvey_dt object}

\item{hhsize}{a household size variable within the hhsurvey_dt object}

\item{adminshp_dt}{an object of class sf, data.table and/or data.frame containing administrative
level boundaries with multipolygons/polygons geometries}

\item{geopolycensus_dt}{an object of class sf, data.table and/or data.frame containing polygon/multipolygon
geometries and geospatial indicators}

\item{geopopvar}{a character string for the population count variable name in the geopolycensus_dt}

\item{crs_set}{an integer list of the coordinate reference systems for of the aforementioned objects
i.e. the CRS for hh_dt, adminshp_dt and geopolycensus_dt in this order}

\item{agr_set}{a character/string list representing attribute-geometry-relationships specified for each
non-geometry attribute column and how it relates to the geometry, and can have one of following values
"constant", "aggregate" and "identity". The default is constant. See details for more. The AGR will listed
in the same order as the crs_set.}

\item{cand_vars}{a character vector of candidate explanatory variables to be included in the model
selection process}

\item{outcome_var}{the dependent variable for small area estimation (typically household per capita expenditure)}

\item{wgt_vartype}{a character string representing the weighting type. The options could be "hh", "pop" i.e.
households vs population weights.}

\item{weight}{the weight variable}

\item{create_dummy}{if TRUE, a dummy variable will be created if dummy_var is specified.}

\item{dummy_var}{a list of variables from which a dummies will be created for each level.}

\item{admin_id}{a character string representing a column vector in hhsurvey_dt for the admin level at which
small area estimates will be computed for the poverty map}

\item{geopoly_id}{a string/character variable representing the polygon ID within geopolycensus_dt}
}
\description{
This function prepares remote sensing data for imputation by checking the admin level shapefiles for geometric
consistency. Prior to the imputation, an algorithm is used to select an appropriate set of variables that best
predicts the outcome variable. The household survey data is used to create a synthetic census of households.
The empirical best predictor is estimated using the household survey data and is used as the basis for imputation
into the synthetic census.
}
