% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_ordernormpl.R
\name{saeplus_ordernormpl}
\alias{saeplus_ordernormpl}
\title{Compute National Poverty Line}
\usage{
saeplus_ordernormpl(
  npl_value = 5006362,
  pcexp = hh.dt$pcexp,
  method = c("inclusion", "interpolation", "limsup")
)
}
\arguments{
\item{npl_value}{a numeric value for the national/international poverty line to be coverted to order norm
scale for imputation}

\item{pcexp}{a numeric vector of per capita expenditure values to be converted to undergo order norm normalization}

\item{method}{a string value indicating the method to be used to convert the npl from currency to normalization.
Three values could be selected "inclusion", "interpolation" and "limsup}
}
\description{
This function computes the poverty line along the order norm scale necessary for EMDI
}
\details{
The "inclusion" method add the poverty line value stipulated to the vector of welfare vector provided. The order
norm value obtained from the normalization process is the converted value
The interpolation method uses the 'pcexp' vector as given and then uses a linear interpolation method to estimate a conversion
The limsup method takes the converted value of the greatest welfare value below the poverty line
}
