% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_ebpreports.R
\name{saeplus_regtable}
\alias{saeplus_regtable}
\title{Create report ready tables from EBP output}
\usage{
saeplus_regtable(
  ebp_obj = ginemdi_model2,
  hhsample_dt = gin_master.dt,
  hhcensus_dt = gin_mastercentroid.dt,
  samplegrid = "id",
  censusgrid = "id"
)
}
\arguments{
\item{ebp_obj}{an object of class "emdi", "ebp"}

\item{hhsample_dt}{household level survey dataset containing grid ID}

\item{hhcensus_dt}{household level population dataset containing grid ID}

\item{samplegrid}{grid ID variable in hhsample_dt}

\item{censusgrid}{grid ID variable in hhcensus_dt}
}
\description{
This function takes the results of the EBP function and creates more outputs to be included in reports
}
