% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeplus_selectmodel.R
\name{saeplus_selectmodel}
\alias{saeplus_selectmodel}
\title{Select Model for Small Area Estimation}
\usage{
saeplus_selectmodel(
  dt,
  outcomevar = "pcexp",
  var_identifier = c("roaddensity_", "count_", "length_", "_pointcount", "bld_",
    "_2018", "_2019"),
  drop_NA_tags = TRUE
)
}
\arguments{
\item{dt}{object of class data.table or data.frame}

\item{outcomevar}{character vector of size 1 naming the outcome variable}

\item{var_identifier}{a character vector of common tags}

\item{drop_NA_tags}{if TRUE, variable names with "NA" in the names are dropped}
}
\value{
a list summarizing the lasso regression output as well as the set of coefficients
}
\description{
This function performs model selection assuming data obtained from Open Street Maps, Google Earth Engine and other sources
from which the SAEplus package functions pull and process geospatial data
}
