% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpopbuilding_pull.R
\name{wpopbuilding_pull}
\alias{wpopbuilding_pull}
\title{Automate the process of pulling building data from the World Pop website}
\usage{
wpopbuilding_pull(iso = "BEN", wpversion = "v2.0", ldrive_dsn = "data-raw")
}
\arguments{
\item{iso}{The standard ISO-3 country code for country of interest}

\item{wpversion}{The World Pop version of interest (use wpopbuilding_vcheck function to check version availability)}

\item{ldrive_dsn}{Destination for the downloaded files}
}
\description{
Automate the process of pulling building data from the World Pop website
}
\details{
For more information on documentation please run wpopbuilding_pull(iso = "XXX", wpversion, ldrive_dsn). This should
download the documentation complete with data file descriptions and citation references
}
