% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpopbuilding_vcheck.R
\name{wpopbuilding_vcheck}
\alias{wpopbuilding_vcheck}
\title{This function returns a datatable of all available versions of World Pop raster
building data in each country with available}
\usage{
wpopbuilding_vcheck()
}
\value{
a data.table object with list and versions of building data
}
\description{
simply run : wpopbuilding_check() in your console.
}
\note{
ALWAYS GIVE AT LEAST 2 MINUTES BETWEEN SUCCESSIVE RUNS OF THIS FUNCTION. THE PREVENTS SUBSEQUENT REQUESTS ON ITS
SERVER WITHIN THAT TIME INTERVAL, SO YOU WILL GET A <NOT SET> FUNCTION ERROR.
}
