% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpopbuilding_vcheck.R
\name{wpopbuilding_vcheck}
\alias{wpopbuilding_vcheck}
\title{This function returns a datatable of all available versions of World Pop raster
building data in each country with available}
\usage{
wpopbuilding_vcheck()
}
\description{
simply run : wpopbuilding_check() in your console
}
